Moved tasks logic to dedicated file
